refactor: 알림 허용 모달 features/notification으로 이동
